fix: fix signed int for mint

+++ packages/whisky/src/builder/core.rs
@@ -563,7 +563,7 @@ impl IMeshTxBuilder for MeshTxBuilder {
         self
     }
 
-    fn mint(&mut self, quantity: u64, policy: &str, name: &str) -> &mut Self {
+    fn mint(&mut self, quantity: i128, policy: &str, name: &str) -> &mut Self {
         if self.mint_item.is_some() {
             self.queue_mint();
         }
